refactor: move utils/types out of components
