--- src/components/calendar/utils.ts
@@ -1,37 +1,0 @@
-//
-// CalendarWeekly Utils
-//
-
-/**
- * Generates the string representation of a date.
- *
- * @param {number} month The integer representation of the month.
- * @param {number} day The integer representation of the day of the month.
- * @param {number} year The integer representation of the year.
- * @returns {string} The date as a string.
- */
-
-export const ALL_MONTHS = [
-  "January",
-  "February",
-  "March",
-  "April",
-  "May",
-  "June",
-  "July",
-  "August",
-  "September",
-  "October",
-  "November",
-  "December",
-]
-
-export const ALL_DAYS_OF_WEEK = [
-  "Sun",
-  "Mon",
-  "Tue",
-  "Wed",
-  "Thur",
-  "Fri",
-  "Sat",
-]
