+++ src/components/storage/StorageTool.tsx
@@ -6,8 +6,8 @@ import { ScrollButton } from "@/components/button/ScrollButton"
 import { ContentSection } from "@/components/ContentSection"
 import { StorageForm } from "@/components/storage/StorageForm"
 import { StorageTable } from "@/components/storage/StorageTable"
-import { TABLE_VISIBILITY } from "@/lib/styles"
-import { SelectedAnswers } from "@/lib/storage-types"
+import { TABLE_VISIBILITY } from "@/styles/visibility"
+import { SelectedAnswers } from "@/types/storage-types"
 import services from "@/content/data/storage-features.json"
 import questions from "@/content/data/storage-questions.json"
