--- src/lib/oscar-types.ts
@@ -1,17 +1,0 @@
-export interface ContentLinks {
-  text: string
-  href: string
-}
-
-export interface OscarSubsection {
-  title: string
-  description: string
-  links?: ContentLinks[]
-}
-
-export interface OscarData {
-  title: string
-  description: string
-  sectionTitle: string
-  subsections: OscarSubsection[]
-}
