+++ src/lib/setup-search-key.ts
