+++ src/types/logo-types.ts
