+++ src/types/storage-types.ts
@@ -1,41 +1,3 @@
-// --- Import Icons ---
-import {
-  FaCamera,
-  FaChalkboardTeacher,
-  FaDollarSign,
-  FaExchangeAlt,
-  FaFile,
-  FaHdd,
-  FaNetworkWired,
-  FaShareSquare,
-  FaShieldAlt,
-  FaSyncAlt,
-  FaTachometerAlt,
-  FaThermometerHalf,
-  FaWifi,
-} from "react-icons/fa"
-import { SiDoi } from "react-icons/si"
-
-// --- Helper Mappings for Icons and Feature Value Colors ---
-export const featureIcons: Record<string, React.ElementType> = {
-  relative_speed: FaTachometerAlt,
-  security: FaShieldAlt,
-  cost: FaDollarSign,
-  sharing: FaShareSquare,
-  capacity: FaHdd,
-  doi_provided: SiDoi,
-  data_protection_snapshots: FaCamera,
-  data_protection_replication: FaSyncAlt,
-  canvas_integration: FaChalkboardTeacher,
-  brown_network_required: FaWifi,
-  access_from_oscar: FaNetworkWired,
-  storage_warmth: FaThermometerHalf,
-  collaborative_edits: FaExchangeAlt,
-  shareable_link: FaShareSquare,
-  max_file_size: FaFile,
-  storage: FaHdd,
-}
-
 export const featureColorMap: Record<string, string> = {
   "0": "text-keppel-700",
   "1": "text-sunglow-400",
